+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+------------
+### Added
+- Hero image support. Set the new `config.heroImage` parameter to embed a hero image into your readme.
+
 [1.0.0] - 2024-10-03
 --------------------
 ### Added